# Refactor to generate a dynamic class for mapping based on properties