--- a/.gitignore
+++ b/.gitignore
@@ -12,3 +12,5 @@ Gemfile.lock
 # rspec failure tracking
 .rspec_status
 .byebug_history
+odata_profile.json
+simple_profile.json