--- a/lib/odata_duty/complex_type.rb
+++ b/lib/odata_duty/complex_type.rb
@@ -65,37 +65,5 @@ module OdataDuty
       @object = object
       @od_context = context
     end
-
-    # def self.to_value(*args)
-    #   new(*args).__to_value
-    # end
-
-    # def __to_value
-    #   MapperBuilder.map(self.class, object)
-
-    #   # self.class.properties.each_with_object({}) do |property, result|
-    #   #   result[property.name] = property.value_from_object(object, od_context)
-    #   # rescue NoMethodError
-    #   #   raise NoImplementationError,
-    #   #         "#{self.class} or #{object.class} do not respond to #{property.name}"
-    #   # end
-    # end
-
-    # def method_missing(method_name, ...)
-    #   if object.respond_to?(method_name)
-    #     # Define the method dynamically
-    #     self.class.define_method(method_name) do
-    #       object.public_send(method_name)
-    #     end
-    #     # Call the newly defined method
-    #     public_send(method_name, ...)
-    #   else
-    #     super
-    #   end
-    # end
-
-    def respond_to_missing?(method_name, include_private = false)
-      object.respond_to?(method_name) || super
-    end
   end
 end