--- a/lib/odata_duty/dynamic_object_wrapper.rb.erb
+++ b/lib/odata_duty/dynamic_object_wrapper.rb.erb
@@ -1,0 +1,62 @@
+class <%= klass_name %> < MapperBuilder
+  def to_h
+    {
+    <% complex_type.properties.each do |property| %>
+      '<%= property.name %>' => <% unless property.nullable? %>not_nullable("<%= property.name %>",<% end %><%= property.name %>_value<% unless property.nullable? %>)<% end %>,
+    <% end %>
+    }
+  end
+
+  <% complex_type.properties.each do |property| %>
+  def <%= property.name %>_value
+    <% if property.calling_method? %>
+    raw_value = calling_methods.fetch(:<%= property.method_name %>).call(obj)
+    <% else %>
+    raw_value = obj.<%= property.method_name %>
+    <% end %>
+    return nil if raw_value.nil?
+
+  <% if property.collection? %>
+    <% if property.scalar? %> 
+      <% if property.boolean? %>
+        raw_value.map{ |rw| confirm_boolean("<%= property.name %>", rw) }
+      <% elsif property.string? %>
+        raw_value.map(&:to_str)
+      <% elsif property.int? %>
+        raw_value.map(&:to_int)
+      <% elsif property.date? %>
+        raw_value.map(&:to_date).map(&:iso8601)
+      <% elsif property.datetime? %>
+        raw_value.map(&:to_datetime).map(&:iso8601)
+      <% else %>
+        raise('<%= property.type %> Coming soon')
+      <% end %>
+    <% else %>
+      mappers['<%= property.name %>'] ||= MapperBuilder.build(complex_types.fetch(:<%= property.name %>), raw_value.first)
+      raw_value.map{ |rw| mappers['<%= property.name %>'].obj_to_hash(rw) }
+    <% end %>
+  <% else %>
+    <% if property.scalar? %> 
+      <% if property.boolean? %>
+        confirm_boolean("<%= property.name %>", raw_value)
+      <% elsif property.string? %>
+        raw_value.to_str
+      <% elsif property.int? %>
+        raw_value.to_int
+      <% elsif property.date? %>
+        raw_value.to_date.iso8601
+      <% elsif property.datetime? %>
+        raw_value.to_datetime.iso8601
+      <% elsif property.enum? %>
+        confirm_one_of("<%= property.name %>", raw_value, %w[<%= property.enum_members.map(&:name).join(' ') %>])
+      <% else %>
+        raise('<%= property.type %> Coming soon')
+      <% end %>
+    <% else %>
+      mappers['<%= property.name %>'] ||= MapperBuilder.build(complex_types.fetch(:<%= property.name %>), raw_value)
+      mappers['<%= property.name %>'].obj_to_hash(raw_value)
+    <% end %>
+  <% end %>
+  end
+  <% end %>
+end