… .byebug_history
… benchmarks/confirm_bool.rb
… benchmarks/flat.rb
… benchmarks/nested.rb
… lib/odata_duty/property.rb
… lib/odata_duty/property/public_send_method.rb
… lib/odata_duty/property/single_prop.rb
… lib/odata_duty/schema_builder/complex_type.rb
… lib/odata_duty/schema_builder/endpoint.rb
… lib/odata_duty/schema_builder/entity_type.rb
… lib/odata_duty/schema_builder/enum_type.rb
… odata_duty.gemspec
… spec/spec_helper.rb