Instead of looping through properties to generate the correct data, this PR instead dynamically generates a mapper class based on the properties. This is a full refactor of the internals but preserving the same behaviour and keeping all the tests passing while reducing overhead caused by odata gem to map data to json.

Below are the benchmarks of doing a simple json map vs using odata

Before Changes:

```bash
ruby 3.3.6 (2024-11-05 revision 75015d4c1f) [x86_64-linux]
Warming up --------------------------------------
              simple     6.000 i/100ms
          odata-duty     1.000 i/100ms
Calculating -------------------------------------
              simple     50.659 (±13.8%) i/s   (19.74 ms/i) -    252.000 in   5.083275s
          odata-duty     13.474 (± 7.4%) i/s   (74.22 ms/i) -     67.000 in   5.006301s

Comparison:
              simple:       50.7 i/s
          odata-duty:       13.5 i/s - 3.76x  slower
```

After Changes:

```bash
ruby 3.3.6 (2024-11-05 revision 75015d4c1f) [x86_64-linux]
Warming up --------------------------------------
              simple     5.000 i/100ms
          odata-duty     3.000 i/100ms
Calculating -------------------------------------
              simple     51.596 (± 5.8%) i/s   (19.38 ms/i) -    260.000 in   5.057515s
          odata-duty     32.711 (± 9.2%) i/s   (30.57 ms/i) -    165.000 in   5.080409s

Comparison:
              simple:       51.6 i/s
          odata-duty:       32.7 i/s - 1.58x  slower
```